+++ src/helper.js
@@ -15,13 +15,15 @@ export function classnames(...arr) {
     return [...classNames].join(' ');
 }
 
+/**
+ * Takes a string and fills placeholders with variables values.
+ * e.g. "I like !{softwareProject}." could become "I like Matrix."
+ */
+const variablePattern = /!{(.+?)}/;
 export function fillInVariables(template, variables) {
-    const variablePattern = /!{(.+?)}/;
     let match = template.match(variablePattern);
     while (match) {
-        console.log(match);
         template = template.slice(0, match.index) + encodeURIComponent(variables[match[1]]) + template.slice(match.index + match[0].length);
-        console.log(template);
         match = template.match(variablePattern);
     }
     return template;